Refactor enpoints into appropriate controllers
